good

+++ prepare-or-build.sh
@@ -4,13 +4,20 @@ if [ -d "out/headless" ]; then
 
   # load pull docker image and run directly
   docker pull microbox/chromium-builder:headless-${VERSION}
-  docker run -it microbox/chromium-builder:headless-${CHROMIUM_VERSION} /bin/bash
-  docker build -t ubuntu-build --build-arg CHROMIUM_VERSION=${VERSION} ubuntu-build
-  export CID=$(docker create ubuntu-build)
-  docker cp $CID:/root/chromium/headless.tar.gz debian
-  docker build -t ${REPO}:${TAG} debian
-  docker push ${REPO}
-  rm -rf out/headless
+  docker run -it microbox/chromium-builder:headless-${CHROMIUM_VERSION} -v out:/root/chromium/src/out timeout 40m ninja -C out/headless headless_shell
+
+  if [ -e "out/headless/headless_shell" ]; then
+
+    tar zcvf debian/headless.tar.gz out/headless/headless_shell out/headless/*.pak
+    docker build -t ${REPO}:${TAG} debian
+    docker push ${REPO}
+    rm -rf out/headless
+  else
+
+    ls -ahl out/headless
+    du -h -d 1 out/headless
+
+  fi;
 
 else
